+++ rust/rsnano/src/ffi/ledger/datastore/lmdb/peer_store.rs
@@ -58,7 +58,7 @@ pub unsafe extern "C" fn rsn_lmdb_peer_store_count(
     handle: *mut LmdbPeerStoreHandle,
     txn: *mut TransactionHandle,
 ) -> usize {
-    (*handle).0.count((*txn).as_txn())
+    (*handle).0.count((*txn).as_txn()) as usize
 }
 
 #[no_mangle]
