+++ rust/rsnano/src/ffi/ledger/datastore/lmdb/pruned_store.rs
@@ -93,7 +93,7 @@ pub unsafe extern "C" fn rsn_lmdb_pruned_store_count(
     handle: *mut LmdbPrunedStoreHandle,
     txn: *mut TransactionHandle,
 ) -> usize {
-    (*handle).0.count((*txn).as_txn())
+    (*handle).0.count((*txn).as_txn()) as usize
 }
 
 #[no_mangle]
