+++ rust/rsnano/src/ledger/datastore/lmdb/account_store.rs
@@ -89,7 +89,7 @@ impl AccountStore for LmdbAccountStore {
         LmdbIteratorImpl::null_iterator()
     }
 
-    fn count(&self, txn: &dyn Transaction) -> usize {
+    fn count(&self, txn: &dyn Transaction) -> u64 {
         count(txn, self.database)
     }
