+++ rust/rsnano/src/ledger/datastore/lmdb/final_vote_store.rs
@@ -109,7 +109,7 @@ impl FinalVoteStore for LmdbFinalVoteStore {
         }
     }
 
-    fn count(&self, txn: &dyn Transaction) -> usize {
+    fn count(&self, txn: &dyn Transaction) -> u64 {
         count(txn, self.database)
     }
