+++ rust/rsnano/src/ledger/datastore/unchecked_store.rs
@@ -12,5 +12,5 @@ pub trait UncheckedStore {
     fn del(&self, txn: &mut dyn WriteTransaction, key: &UncheckedKey);
     fn begin(&self, txn: &dyn Transaction) -> UncheckedIterator;
     fn lower_bound(&self, txn: &dyn Transaction, key: &UncheckedKey) -> UncheckedIterator;
-    fn count(&self, txn: &dyn Transaction) -> usize;
+    fn count(&self, txn: &dyn Transaction) -> u64;
 }
